Added categories and documentation links to the report messages.
Introduced a new class CReportMessage to hold all information of a report message,
including the category and the documentation link. Updated the macros for logging
such that it is possible to pass a category and a documentation link. Also updated
the formatting of the Json output, such that all information concerning a report
message is logged. The order of the messages is now also preserved.