Add test checking if errors end up in log file.
There was some doubt expressed in the merge request comments as to
whether messages reported with CREPORT_ERROR_NODOC are also logged to
the normal log file when reporting is turned off.

See https://github.com/KNMI/adaguc-server/pull/66/files#r199443944 and
https://github.com/KNMI/adaguc-server/pull/66/files#r199445156
This test doesn't test that particular behaviour. It tests whether
errors end up in the logfile when reporting is turned on, but this test
can serve as basis for testing the case when reporting is turned off.